deploy/requirements.txt
google-adk>=1.31.0
python-dotenv
uvicorn[standard]
aiosqlite
asyncpg
